Extract components to separate files
